shoe-store-app/requirements.txt
Pillow>=9.0.0
flake8>=6.0.0
